+++ src/main/java/com/yohanii/lostandfound/component/init/InitService.java
@@ -10,7 +10,6 @@ import com.yohanii.lostandfound.component.crud.repository.PostRepository;
 import com.yohanii.lostandfound.component.crud.service.PostService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -35,8 +34,6 @@ public class InitService {
     private final PostService postService;
     private final PostRepository postRepository;
 
-    public static final int INIT_MEMBER_COUNT = 1000;
-    public static final long INIT_POST_COUNT = 1000;
     public static final String INIT_IMAGE_PATH = "src/main/resources/static/img/item/";
     public static final Map<String, Integer> itemImages = Map.of(
             "airpot.jpeg", 3,
@@ -70,14 +67,14 @@ public class InitService {
     }
 
     @Transactional
-    public int fillMembers() {
+    public int fillMembers(int count) {
 
-        if (memberRepository.count() >= INIT_MEMBER_COUNT) {
+        if (memberRepository.count() >= count) {
             return 0;
         }
 
         return memberRepository.saveAll(
-                        IntStream.range(0, INIT_MEMBER_COUNT)
+                        IntStream.range(0, count)
                                 .mapToObj(index -> {
                                     LocalDateTime now = LocalDateTime.now();
 
@@ -95,9 +92,9 @@ public class InitService {
                 .size();
     }
 
-    public long fillPosts() {
+    public long fillPosts(int count) {
 
-        if (postRepository.count() >= INIT_POST_COUNT) {
+        if (postRepository.count() >= count) {
             return 0L;
         }
 
@@ -163,7 +160,7 @@ public class InitService {
 
         for (int index = 1; index <= count; index++) {
             String name = split[0] + index + "." + split[1];
-            result.add(new MockMultipartFile(name, name, "text/plain", contents.getOrDefault(name, null)));
+//            result.add(new MockMultipartFile(name, name, "text/plain", contents.getOrDefault(name, null)));
         }
 
         return result;