+++ src/test/java/com/yohanii/lostandfound/component/crud/entity/PostTest.java
@@ -1,8 +1,7 @@
 package com.yohanii.lostandfound.component.crud.entity;
 
-import com.yohanii.lostandfound.component.crud.entity.Post;
-import com.yohanii.lostandfound.component.crud.entity.PostType;
 import com.yohanii.lostandfound.component.crud.dto.post.PostEditRequestDto;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -10,8 +9,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 class PostTest {
 
     @Test
+    @DisplayName("포스트 업데이트 성공")
     void updatePost() {
-        Post testPost = Post.builder()
+
+        //given
+        Post post = Post.builder()
                 .title("post title")
                 .content("post content")
                 .type(PostType.LOST)
@@ -22,10 +24,12 @@ class PostTest {
         dto.setContent("edit content");
         dto.setType(PostType.FOUND);
 
-        testPost.updatePost(dto);
+        //when
+        post.updatePost(dto);
 
-        assertThat(testPost.getTitle()).isEqualTo(dto.getTitle());
-        assertThat(testPost.getContent()).isEqualTo(dto.getContent());
-        assertThat(testPost.getType()).isEqualTo(dto.getType());
+        //then
+        assertThat(post.getTitle()).isEqualTo(dto.getTitle());
+        assertThat(post.getContent()).isEqualTo(dto.getContent());
+        assertThat(post.getType()).isEqualTo(dto.getType());
     }
 }