+++ src/test/java/com/yohanii/lostandfound/component/crud/service/ChattingServiceTest.java
@@ -9,58 +9,93 @@ import com.yohanii.lostandfound.component.chatting.repository.RoomRepository;
 import com.yohanii.lostandfound.component.chatting.service.ChattingService;
 import com.yohanii.lostandfound.component.crud.entity.Member;
 import com.yohanii.lostandfound.component.crud.repository.MemberRepository;
-import jakarta.transaction.Transactional;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.BDDMockito.given;
 
-@SpringBootTest
-@Transactional
+@ExtendWith(MockitoExtension.class)
 class ChattingServiceTest {
 
-    @Autowired
+    @InjectMocks
     ChattingService chattingService;
-    @Autowired
+    @Mock
     MemberRepository memberRepository;
-    @Autowired
+    @Mock
     RoomRepository roomRepository;
-    @Autowired
+    @Mock
     ChattingRepository chattingRepository;
 
     @Test
+    @DisplayName("실행 시 dto의 memberId, roomId에 해당하는 객체를 담은 Chatting 객체를 반환해야한다.")
     void saveChatting() {
-        Member member = Member.builder().build();
-        Long savedMemberId = memberRepository.save(member).getId();
-        Room room = Room.builder().build();
-        Long savedRoomId = roomRepository.save(room).getId();
 
+        //given
+        Member member = Member.builder()
+                .id(1L)
+                .build();
+        Room room = Room.builder()
+                .id(1L)
+                .build();
+        ChattingMessageDto dto = new ChattingMessageDto(member.getId(), room.getId(), ChattingType.ENTER, "입장했습니다.");
 
-        ChattingMessageDto dto = new ChattingMessageDto(savedMemberId, savedRoomId, ChattingType.ENTER, "입장했습니다.");
-        Long savedChattingId = chattingService.saveChatting(dto);
+        given(memberRepository.findById(member.getId())).willReturn(Optional.of(member));
+        given(roomRepository.findById(room.getId())).willReturn(Optional.of(room));
+        given(chattingRepository.save(any(Chatting.class)))
+                .willReturn(Chatting.builder()
+                        .member(member)
+                        .room(room)
+                        .type(dto.getType())
+                        .content(dto.getContent())
+                        .build());
 
-        Chatting findChatting = chattingRepository.findById(savedChattingId).orElse(null);
+        //when
+        Chatting result = chattingService.saveChatting(dto);
 
-        assertThat(findChatting.getMember()).isEqualTo(member);
-        assertThat(findChatting.getRoom()).isEqualTo(room);
-        assertThat(findChatting.getType()).isEqualTo(dto.getType());
-        assertThat(findChatting.getContent()).isEqualTo(dto.getContent());
+        //then
+        assertThat(result.getMember()).isEqualTo(member);
+        assertThat(result.getRoom()).isEqualTo(room);
+        assertThat(result.getType()).isEqualTo(dto.getType());
+        assertThat(result.getContent()).isEqualTo(dto.getContent());
     }
 
     @Test
-    void saveChatting시_room의_updatedPost_update() {
-        Member member = Member.builder().build();
-        Long savedMemberId = memberRepository.save(member).getId();
-        Room room = Room.builder().build();
-        Long savedRoomId = roomRepository.save(room).getId();
-        ChattingMessageDto dto = new ChattingMessageDto(savedMemberId, savedRoomId, ChattingType.ENTER, "입장했습니다.");
+    @DisplayName("dto의 memberId에 해당하는 member 없을 시, NoSuchElementException 발생")
+    void saveChatting_해당하는_유저_없음() {
 
-        Long savedChattingId = chattingService.saveChatting(dto);
+        //given
+        ChattingMessageDto dto = new ChattingMessageDto(1L, 1L, ChattingType.ENTER, "입장했습니다.");
 
-        Chatting findChatting = chattingRepository.findById(savedChattingId).orElse(null);
-        Room chattingRoom = findChatting.getRoom();
+        given(memberRepository.findById(anyLong())).willReturn(Optional.ofNullable(null));
 
-        assertThat(chattingRoom.getUpdatedTime()).isEqualTo(findChatting.getCreatedTime());
+        //when -> then
+        assertThatThrownBy(() -> chattingService.saveChatting(dto))
+                .isInstanceOf(NoSuchElementException.class);
+    }
+
+    @Test
+    @DisplayName("dto의 roomId에 해당하는 room 없을 시, NoSuchElementException 발생")
+    void saveChatting_해당하는_채팅방_없음() {
+
+        //given
+        ChattingMessageDto dto = new ChattingMessageDto(1L, 1L, ChattingType.ENTER, "입장했습니다.");
+
+        given(memberRepository.findById(anyLong())).willReturn(Optional.ofNullable(Member.builder().build()));
+        given(roomRepository.findById(anyLong())).willReturn(Optional.ofNullable(null));
+
+        //when -> then
+        assertThatThrownBy(() -> chattingService.saveChatting(dto))
+                .isInstanceOf(NoSuchElementException.class);
     }
 }