+++ src/test/java/com/yohanii/lostandfound/component/crud/service/ImageStoreServiceTest.java
@@ -1,19 +1,20 @@
 package com.yohanii.lostandfound.component.crud.service;
 
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.yohanii.lostandfound.component.crud.dto.image.ItemImagesSaveDto;
 import com.yohanii.lostandfound.component.crud.dto.image.ProfileImageSaveDto;
 import com.yohanii.lostandfound.component.crud.entity.Image;
-import com.yohanii.lostandfound.component.crud.entity.ImageType;
 import com.yohanii.lostandfound.component.crud.entity.Item;
 import com.yohanii.lostandfound.component.crud.entity.Member;
-import com.yohanii.lostandfound.component.crud.repository.ItemRepository;
-import com.yohanii.lostandfound.component.crud.repository.MemberRepository;
-import jakarta.transaction.Transactional;
+import com.yohanii.lostandfound.component.crud.repository.ImageRepository;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -21,67 +22,74 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.BDDMockito.*;
 
-@SpringBootTest
-@Transactional
+@ExtendWith(MockitoExtension.class)
 class ImageStoreServiceTest {
 
-    @Autowired
+    @InjectMocks
     ImageStoreService imageStoreService;
-    @Autowired
-    MemberRepository memberRepository;
-    @Autowired
-    ItemRepository itemRepository;
+    @Mock
+    ImageRepository imageRepository;
+    @Mock
+    AmazonS3 s3Client;
 
     @Test
+    @DisplayName("프로필 이미지 저장 시 이미지 빈 객체 예외 처리")
     void saveImage_profileImage_isEmpty() {
+
+        //given
         ProfileImageSaveDto dto = new ProfileImageSaveDto(null, null);
 
+        //when -> then
         assertThatThrownBy(() -> imageStoreService.saveImage(dto))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("ProfileImage is Empty");
     }
 
     @Test
+    @DisplayName("프로필 이미지 저장 성공")
     void saveImage() throws IOException {
+
+        //given
         Member member = Member.builder().build();
-        memberRepository.save(member);
         MockMultipartFile file = new MockMultipartFile("file", "test.jpeg", "", InputStream.nullInputStream());
+
         ProfileImageSaveDto dto = new ProfileImageSaveDto(member, file);
 
+        Image resultImage = Image.builder().build();
+        given(imageRepository.save(any(Image.class))).willReturn(resultImage);
+
+        //when
         Image savedImage = imageStoreService.saveImage(dto);
 
-        assertThat(savedImage.getId()).isNotNull();
-        assertThat(savedImage.getMember()).isEqualTo(member);
-        assertThat(savedImage.getItem()).isNull();
-        assertThat(savedImage.getType()).isEqualTo(ImageType.MEMBER);
-        assertThat(savedImage.getUploadFileName()).isEqualTo(file.getOriginalFilename());
-        assertThat(savedImage.getStoreFileName()).isNotBlank();
+        //then
+        assertThat(resultImage).isEqualTo(savedImage);
+        then(s3Client).should().putObject(any(), anyString(), any(InputStream.class), any(ObjectMetadata.class));
+        then(imageRepository).should().save(any(Image.class));
     }
 
     @Test
+    @DisplayName("아이템 이미지 저장 성공")
     void saveImages() throws IOException {
+
+        //given
         Item item = Item.builder().build();
-        itemRepository.save(item);
         MockMultipartFile file1 = new MockMultipartFile("file1", "test1.jpeg", "", InputStream.nullInputStream());
         MockMultipartFile file2 = new MockMultipartFile("file2", "test2.jpeg", "", InputStream.nullInputStream());
-        List<MultipartFile> files = List.of(file1, file2);
-        ItemImagesSaveDto dto = new ItemImagesSaveDto(item, files);
 
+        ItemImagesSaveDto dto = new ItemImagesSaveDto(item, List.of(file1, file2));
+
+        Image resultImage = Image.builder().build();
+        given(imageRepository.save(any(Image.class))).willReturn(resultImage);
+
+        //when
         List<Image> savedImages = imageStoreService.saveImages(dto);
 
+        //then
+        assertThat(savedImages).contains(resultImage);
         assertThat(savedImages.size()).isEqualTo(2);
-        assertThat(savedImages.get(0).getId()).isNotNull();
-        assertThat(savedImages.get(0).getItem()).isEqualTo(item);
-        assertThat(savedImages.get(0).getMember()).isNull();
-        assertThat(savedImages.get(0).getType()).isEqualTo(ImageType.ITEM);
-        assertThat(savedImages.get(0).getUploadFileName()).isEqualTo(file1.getOriginalFilename());
-        assertThat(savedImages.get(0).getStoreFileName()).isNotBlank();
-        assertThat(savedImages.get(1).getId()).isNotNull();
-        assertThat(savedImages.get(1).getItem()).isEqualTo(item);
-        assertThat(savedImages.get(1).getMember()).isNull();
-        assertThat(savedImages.get(1).getType()).isEqualTo(ImageType.ITEM);
-        assertThat(savedImages.get(1).getUploadFileName()).isEqualTo(file2.getOriginalFilename());
-        assertThat(savedImages.get(1).getStoreFileName()).isNotBlank();
+        then(s3Client).should(times(2)).putObject(any(), anyString(), any(InputStream.class), any(ObjectMetadata.class));
+        then(imageRepository).should(times(2)).save(any(Image.class));
     }
 }