+++ src/test/java/com/yohanii/lostandfound/component/crud/service/MemberServiceTest.java
@@ -1,10 +1,9 @@
 package com.yohanii.lostandfound.component.crud.service;
 
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
 
-@SpringBootTest
-@Transactional
+@ExtendWith(MockitoExtension.class)
 class MemberServiceTest {
 
 }