+++ src/test/java/com/yohanii/lostandfound/component/crud/service/RoomServiceTest.java
@@ -8,127 +8,132 @@ import com.yohanii.lostandfound.component.crud.entity.Member;
 import com.yohanii.lostandfound.component.crud.entity.Post;
 import com.yohanii.lostandfound.component.crud.repository.MemberRepository;
 import com.yohanii.lostandfound.component.crud.repository.PostRepository;
-import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.List;
+import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
 
-@SpringBootTest
-@Transactional
+@ExtendWith(MockitoExtension.class)
 class RoomServiceTest {
 
-    @Autowired
+    @InjectMocks
     RoomService roomService;
-    @Autowired
+    @Mock
     RoomRepository roomRepository;
-    @Autowired
+    @Mock
     MemberRepository memberRepository;
-    @Autowired
+    @Mock
     PostRepository postRepository;
 
     @Test
+    @DisplayName("채팅방 생성 성공")
     void createRoom() {
-        Member member = Member.builder().build();
-        Long savedMemberId = memberRepository.save(member).getId();
-        Post post = Post.builder().build();
-        Long savedPostId = postRepository.save(post).getId();
 
-        RoomSaveRequestDto dto = new RoomSaveRequestDto();
-        dto.setMemberId(savedMemberId);
-        dto.setPartnerId(100L);
-        dto.setPostId(savedPostId);
-        Long savedRoomId = roomService.createRoom(dto);
-
-        Room savedRoom = roomRepository.findById(savedRoomId).orElse(null);
-
-        assertThat(savedRoom.getMember()).isEqualTo(member);
-        assertThat(savedRoom.getPost()).isEqualTo(post);
-        assertThat(savedRoom.getPartnerId()).isEqualTo(dto.getPartnerId());
-        assertThat(savedRoom.getStoreRoomName()).isNotBlank();
-    }
-
-    @Test
-    void getStoreRoomNameById() {
-        Room room = Room.builder()
-                .storeRoomName("testStoreRoomName")
+        //given
+        Member member = Member.builder()
+                .id(1L)
                 .build();
-        Long savedRoomId = roomRepository.save(room).getId();
+        Post post = Post.builder()
+                .id(1L)
+                .build();
 
-        String storeRoomName = roomService.getStoreRoomNameById(savedRoomId);
+        RoomSaveRequestDto dto = new RoomSaveRequestDto(member.getId(), 2L, post.getId());
 
-        assertThat(storeRoomName).isEqualTo(room.getStoreRoomName());
+        given(memberRepository.findById(member.getId())).willReturn(Optional.of(member));
+        given(postRepository.findById(post.getId())).willReturn(Optional.of(post));
+
+        Room roomResult = Room.builder()
+                .member(member)
+                .post(post)
+                .partnerId(dto.getPartnerId())
+                .build();
+        given(roomRepository.save(any(Room.class))).willReturn(roomResult);
+
+        //when
+        Room savedRoom = roomService.createRoom(dto);
+
+        //then
+        assertThat(savedRoom).isEqualTo(roomResult);
     }
 
     @Test
+    @DisplayName("memberId에 해당하는 채팅방을 반환한다.")
     void findRoomByMemberId() {
-        Member member = Member.builder().build();
-        Member otherMember1 = Member.builder().build();
-        Member otherMember2 = Member.builder().build();
-        memberRepository.save(member);
-        memberRepository.save(otherMember1);
-        memberRepository.save(otherMember2);
+
+        //given
+        Member member = Member.builder().id(1L).build();
+        Member partner = Member.builder().id(2L).build();
 
         Room saveRoom1 = Room.builder()
                 .member(member)
-                .partnerId(10001L)
+                .partnerId(partner.getId())
                 .build();
-
         Room saveRoom2 = Room.builder()
-                .member(otherMember1)
+                .member(partner)
                 .partnerId(member.getId())
                 .build();
 
-        Room saveRoom3 = Room.builder()
-                .member(otherMember2)
-                .partnerId(member.getId())
-                .build();
-        roomRepository.save(saveRoom1);
-        roomRepository.save(saveRoom2);
-        roomRepository.save(saveRoom3);
+        given(roomRepository.findAllByMemberId(member.getId())).willReturn(List.of(saveRoom1));
+        given(roomRepository.findAllByPartnerId(member.getId())).willReturn(List.of(saveRoom2));
 
+        //when
         List<Room> findRooms = roomService.findRoomByMemberId(member.getId());
-        assertThat(findRooms.size()).isEqualTo(3);
-        assertThat(findRooms).contains(saveRoom1, saveRoom2, saveRoom3);
+
+        //then
+        assertThat(findRooms.size()).isEqualTo(2);
+        assertThat(findRooms).contains(saveRoom1, saveRoom2);
     }
 
     @Test
     @DisplayName("memberId, partnerId가 일치하는 room을 반환한다.")
     void findByMemberIdAndPartnerId() {
-        Member member = Member.builder().build();
-        Member savedMember = memberRepository.save(member);
 
-        Room room = Room.builder()
-                .member(savedMember)
-                .partnerId(-1L)
+        //given
+        Member member = Member.builder().id(1L).build();
+        Member partner = Member.builder().id(2L).build();
+
+        Room saveRoom = Room.builder()
+                .member(member)
+                .partnerId(partner.getId())
                 .build();
 
-        Room savedRoom = roomRepository.save(room);
+        given(roomRepository.findByMemberIdAndPartnerId(member.getId(), partner.getId())).willReturn(Optional.of(saveRoom));
 
-        Room result = roomService.findRoomByIds(savedMember.getId(), room.getPartnerId()).orElse(null);
+        //when
+        Room result = roomService.findRoomByIds(member.getId(), partner.getId()).orElse(null);
 
-        assertThat(result).isEqualTo(savedRoom);
+        //then
+        assertThat(result).isEqualTo(saveRoom);
     }
 
     @Test
     @DisplayName("memberId, partnerId 서로 바뀐 상태에서도 해당 room을 반환한다.")
     void findByMemberIdAndPartnerId_reverse() {
-        Member member = Member.builder().build();
-        Member savedMember = memberRepository.save(member);
 
-        Room room = Room.builder()
-                .member(savedMember)
-                .partnerId(-1L)
+        //given
+        Member member = Member.builder().id(1L).build();
+        Member partner = Member.builder().id(2L).build();
+
+        Room saveRoom = Room.builder()
+                .member(member)
+                .partnerId(partner.getId())
                 .build();
 
-        Room savedRoom = roomRepository.save(room);
+        given(roomRepository.findByMemberIdAndPartnerId(partner.getId(), member.getId())).willReturn(Optional.of(saveRoom));
 
-        Room result = roomService.findRoomByIds(room.getPartnerId(), savedMember.getId()).orElse(null);
+        //when
+        Room result = roomService.findRoomByIds(partner.getId(), member.getId()).orElse(null);
 
-        assertThat(result).isEqualTo(savedRoom);
+        //then
+        assertThat(result).isEqualTo(saveRoom);
     }
 }